+++ src/main/java/com/example/demo/controller/WingRoomController.java
@@ -11,7 +11,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.demo.Model.Student;
+import com.example.demo.Model.User;
 import com.example.demo.Model.Wing_rooms;
+import com.example.demo.services.UserService;
 import com.example.demo.services.Wing_Room_Service;
 
 @RestController
@@ -20,39 +23,98 @@ public class WingRoomController {
 			@Autowired
 			private Wing_Room_Service wing_Room_Service;
 		
-	
+			@Autowired
+			private UserService userService;
 	
 			@GetMapping("Room")
-			public List<Wing_rooms> getRoom() {
+			public List<Wing_rooms> getRoom(@RequestParam String username , @RequestParam String password) {
+				
+				User user = userService.cheakUseer(username, password);
 				
 				
-				return wing_Room_Service.getAllWings();
+				if(user!=null) {
+					String role = user.getRole();
+				
+				if(role.equals("admin")||role.equals("clerk")||role.equals("viewer") )	
+				{
+					
+					 return wing_Room_Service.getAllWings();
+				}
+				
+				else return null;
+			}
+			else return null;
+			
+				
 							
 			}
 			
 			@PostMapping("Room")
-			public String insertRoom(@RequestBody Wing_rooms room) {
+			public String insertRoom(@RequestBody Wing_rooms room, @RequestParam String username , @RequestParam String password) {
 				
+				User user = userService.cheakUseer(username, password);
 				
-				 return "Inserted : "+wing_Room_Service.insertRoom( room);
+				if(user!=null) {
+					String role = user.getRole();
+					
+					if(role.equals("admin")||role.equals("clerk"))
+						
+					{
+						
+						  return "Inserted : "+wing_Room_Service.insertRoom( room);
+					}
+					
+					else return "User does not have authority to insert ...";
+				}
+				else return "User not found ... ";
+				
 			}
 			
 			
 			@PutMapping("Room")
-			public void updateRoom( @RequestBody Wing_rooms room ) {
+			public String updateRoom( @RequestBody Wing_rooms room,@RequestParam String username , @RequestParam String password ) {
 		  
-				   
-				wing_Room_Service.update(room);
+				User user = userService.cheakUseer(username, password);
+				
+						
+				if(user!=null) {
+					String role = user.getRole();
+					
+					if(role.equals("admin")||role.equals("clerk"))
+						
+					{
+			         	return "Updated :"+wing_Room_Service.update(room);
+					}
+					
+					else return "User does not have authority to update ...";
+				}
+				else return "User not found ...";
+				
 				
 				
 			}
 			
 			@DeleteMapping("Room")
-			public String delete(@RequestParam String id) {
+			public String delete(@RequestParam String id , @RequestParam String username , @RequestParam String password) {
 				
-				wing_Room_Service.delete(id);
 				
-				return "Deleted   ";
+	                User user = userService.cheakUseer(username, password);
+		
+				if(user!=null) {
+					String role = user.getRole();
+					
+					if(role.equals("admin"))
+						
+					{
+						
+						wing_Room_Service.delete(id);
+						return "Deleted   ";
+					}
+					
+					else return "User does not have authority to delete ...";
+				}
+				else return "User not found ...";
+			
 				
 			}
